requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.7.3
    # via
    #   -r requirements.in
    #   dblpy
    #   discord.py
    #   gcsfs
amqp==2.6.1
    # via kombu
argh==0.26.2
    # via watchdog
astroid==2.4.2
    # via pylint
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via
    #   aiohttp
    #   pytest
babelfish==0.5.5
    # via guessit
billiard==3.6.3.0
    # via celery
cachetools==4.2.1
    # via google-auth
celery==4.4.7
    # via
    #   -r requirements.in
    #   flower
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
dataclasses-json==0.5.2
    # via -r requirements.in
dblpy==0.4.0
    # via -r requirements.in
decorator==4.4.2
    # via
    #   gcsfs
    #   moviepy
discord.py==1.6.0
    # via
    #   -r requirements.in
    #   dblpy
elastic-app-search==7.10.0
    # via -r requirements.in
https://github.com/hackslashX/flower/zipball/master
    # via -r requirements.in
fsspec==0.8.5
    # via
    #   -r requirements.in
    #   gcsfs
gcsfs==0.7.1
    # via -r requirements.in
google-auth-oauthlib==0.4.2
    # via gcsfs
google-auth==1.24.0
    # via
    #   gcsfs
    #   google-auth-oauthlib
guessit==3.2.0
    # via -r requirements.in
humanize==3.2.0
    # via flower
idna==2.10
    # via
    #   requests
    #   yarl
imageio-ffmpeg==0.4.3
    # via moviepy
imageio==2.9.0
    # via moviepy
inflect==5.0.2
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
invoke==1.5.0
    # via -r requirements.in
isort==5.7.0
    # via pylint
kombu==4.6.11
    # via celery
lazy-object-proxy==1.4.3
    # via astroid
loguru==0.5.3
    # via -r requirements.in
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow==3.10.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
mccabe==0.6.1
    # via pylint
meilisearch==0.14.0
    # via -r requirements.in
moviepy==1.0.3
    # via -r requirements.in
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via typing-inspect
numpy==1.19.5
    # via
    #   imageio
    #   moviepy
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.8
    # via pytest
pillow==8.1.0
    # via imageio
pluggy==0.13.1
    # via pytest
proglog==0.1.9
    # via moviepy
prometheus_client==0.8.0
    # via flower
py==1.10.0
    # via pytest
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyjwt==1.7.1
    # via elastic-app-search
pylint==2.6.0
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytest==6.2.2
    # via -r requirements.in
python-dateutil==2.8.1
    # via guessit
pytz==2020.5
    # via
    #   celery
    #   flower
pyyaml==5.4.1
    # via watchdog
ratelimiter==1.2.0.post0
    # via dblpy
rebulk==3.0.1
    # via guessit
redis==3.5.3
    # via -r requirements.in
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   -r requirements.in
    #   elastic-app-search
    #   gcsfs
    #   meilisearch
    #   moviepy
    #   requests-oauthlib
rsa==4.7
    # via google-auth
sentry-sdk==0.19.5
    # via -r requirements.in
sh==1.14.1
    # via -r requirements.in
six==1.15.0
    # via
    #   astroid
    #   google-auth
    #   python-dateutil
srt==3.4.1
    # via -r requirements.in
stringcase==1.2.0
    # via dataclasses-json
toml==0.10.2
    # via
    #   pylint
    #   pytest
tornado==6.1
    # via flower
tqdm==4.56.0
    # via
    #   moviepy
    #   proglog
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   typing-inspect
typing-inspect==0.6.0
    # via dataclasses-json
urllib3==1.26.3
    # via
    #   requests
    #   sentry-sdk
vine==1.3.0
    # via
    #   amqp
    #   celery
watchdog[watchmedo]==1.0.2
    # via -r requirements.in
wrapt==1.12.1
    # via astroid
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
